Assignment2/requirements.txt
apache-airflow
clickhouse-connect
